rename DB models acccordingly
