+++ backend/src/models/userModel.ts
@@ -1,7 +1,7 @@
 import { ResultSetHeader } from "mysql2";
 import { PoolConnection } from "mysql2/promise";
 import sharp from "sharp";
-import { UserDB } from "../common/databaseModels.js";
+import { UserDB } from "../common/generated.js";
 import connectionPool from "../config/database.js";
 import ImageModel from "./imageModel.js";
